Add `mix tailwind.install` to setup command
If you do not have a binary of tailwind available, the command will get
it properly linked up and downloaded for you for the appropriate
architecture.

I ran into this while forking this project to work on some campaign
financing analysis tooling for NYS and figured I'd just upstream it
since many others may not have this binary available if learning from
newer posts like this.